backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
aiohttp
google-generativeai
tavily-python
pydantic
pydantic-settings
beautifulsoup4
requests
html2text
sentence-transformers
chromadb
numpy
qdrant-client[fastembed]
firebase-admin
python-multipart